% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{add_tiny}
\alias{add_tiny}
\title{Add a tiny value to the elements of a vector that are equal to zero.}
\usage{
add_tiny(x)
}
\arguments{
\item{x}{Vector of values.}
}
\value{
A vector with zeros replaced by a tiny value.
}
\description{
Add a tiny value to the elements of a vector that are equal to zero.
}
\keyword{internal}
