% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_dad_data}
\alias{check_dad_data}
\title{check validity of DAD data}
\usage{
check_dad_data(dad)
}
\arguments{
\item{dad}{}
}
\value{
Logical value indicating whether the data is valid
}
\description{
check validity of DAD data
}
