% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_date_range_fit}
\alias{check_date_range_fit}
\title{Check if the date range is valid}
\usage{
check_date_range_fit(d)
}
\arguments{
\item{date.fit.range}{Vector of two dates.
The first date must be earlier than the second date.}
}
\description{
Check if the date range is valid
}
