% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_expl_data}
\alias{check_expl_data}
\title{Check validity of explanatory data}
\usage{
check_expl_data(x)
}
\arguments{
\item{x}{a data frame containing explanatory data}
}
\value{
Logical value indicating whether the data is valid
}
\description{
Check validity of explanatory data
}
