% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_fitted_model}
\alias{check_fitted_model}
\title{Check the validity of a fitted model}
\usage{
check_fitted_model(fitted.model)
}
\arguments{
\item{fitted.model}{A list containing the fitted model
and its associated data.}
}
\value{
Logical value indicating whether the fitted model is valid.
}
\description{
Check the validity of a fitted model
}
