% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_newdata_expl}
\alias{check_newdata_expl}
\title{Check the compatibility of the new exploratory data
and the fitted model used for nowcast.}
\usage{
check_newdata_expl(newdata.expl, fitted.model)
}
\arguments{
\item{newdata.expl}{Dataframe of new explanatory data}

\item{fitted.model}{List. Fitted model.}
}
\value{
Logical value indicating whether the new explanatory data is valid.
}
\description{
Check the compatibility of the new exploratory data
and the fitted model used for nowcast.
}
