% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_varname}
\alias{check_varname}
\title{Check if the variable name is present in the dataframe.}
\usage{
check_varname(data, varname)
}
\arguments{
\item{data}{Dataframe of  data}

\item{varname}{String. Name of the variable in data to use as a variable.}
}
\value{
Logical value indicating whether the variable name is valid
}
\description{
Check if the variable name is present in the dataframe.
}
\keyword{internal}
