% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_varname_dad}
\alias{check_varname_dad}
\title{Check if the variable name for DAD data is valid}
\usage{
check_varname_dad(varname.dad)
}
\arguments{
\item{varname.dad}{String. Must be either "count" or "percapita".}
}
\value{
Logical value indicating whether the variable name is valid
}
\description{
Check if the variable name for DAD data is valid
}
