% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_varname_expl}
\alias{check_varname_expl}
\title{Check if the variable name for explanatory data is valid}
\usage{
check_varname_expl(data.expl, varname.expl)
}
\arguments{
\item{data.expl}{Dataframe of explanatory data}

\item{varname.expl}{String. Name of the variable in data.expl to use as an explanatory variable.}
}
\value{
Logical value indicating whether the variable name is valid
}
\description{
Check if the variable name for explanatory data is valid
}
\keyword{internal}
