% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fake_dad}
\alias{fake_dad}
\title{Create fake DAD data for testing purposes.}
\usage{
fake_dad()
}
\value{
A dataframe.
}
\description{
Create fake DAD data for testing purposes.
}
\examples{

data.dad = fake_dad()
head(data.dad)

}
