% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fake_expl}
\alias{fake_expl}
\title{Create fake explanatory data from existing DAD for testing purposes.}
\usage{
fake_expl(dad)
}
\arguments{
\item{dad}{Dataframe of DAD data.}
}
\value{
A dataframe with explanatory data.
}
\description{
Create fake explanatory data from existing DAD for testing purposes.
}
\examples{
data.dad = fake_dad()
data.expl = fake_expl(data.dad)

}
