% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_geo}
\alias{get_geo}
\title{Extract the geography of a fitted model.}
\usage{
get_geo(fitted.model)
}
\arguments{
\item{fitted.model}{List. Fitted model.}
}
\value{
String. geography of the fitted model.
}
\description{
Extract the geography of a fitted model.
}
