% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_virus}
\alias{get_virus}
\title{Extract the virus of a fitted model.}
\usage{
get_virus(fitted.model)
}
\arguments{
\item{fitted.model}{List. Fitted model.}
}
\value{
String. Virus of the fitted model.
}
\description{
Extract the virus of a fitted model.
}
