% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{id_model}
\alias{id_model}
\title{Create a unique ID for a fitted model.}
\usage{
id_model(fitted.model)
}
\arguments{
\item{fitted.model}{List. Fitted model object.}
}
\value{
String of the Id.
}
\description{
Create a unique ID for a fitted model.
}
\keyword{internal}
