% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{join_dad_expl}
\alias{join_dad_expl}
\title{Join DAD and explanatory data}
\usage{
join_dad_expl(data.dad, data.expl, varname.dad, varname.expl)
}
\arguments{
\item{data.dad}{Dataframe of DAD data}

\item{data.expl}{Dataframe of explanatory data}

\item{varname.dad}{String. Name of the variable in data.dad to fit the model to.}

\item{varname.expl}{String. Name of the variable in data.expl to use as an explanatory variable.}
}
\value{
Dataframe
}
\description{
Join DAD and explanatory data
}
\keyword{internal}
