% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcast.R
\name{nowcast}
\alias{nowcast}
\title{Nowcast DAD data using fresh explanatory data.}
\usage{
nowcast(fitted.model, newdata.expl, ci = 0.95, floor.zero = TRUE)
}
\arguments{
\item{fitted.model}{List. Fitted model as created by the function \code{fit_model()}.}

\item{newdata.expl}{Dataframe of new explanatory data. Must contain colums named
\code{date, virus, geo} and a column name matching \code{fitted.model$varname.expl}.}

\item{ci}{Numerical. Confidence interval of the prediction. Must be between 0 and 1.}

\item{floor.zero}{Logical. If `TRUE` nowcast will be floored at 0 when estimates are negatives.}
}
\value{
A list:
\itemize{
    \item \code{nowcast}: Dataframe of nowcasted DAD values.
    \item \code{fitted.model}: Fitted model used to nowcast.
    \item \code{newdata.expl}: New explanatory data used to nowcast.
}
}
\description{
Nowcast DAD data using fresh explanatory data.
}
