% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_fitted_model}
\alias{plot_fitted_model}
\title{Plot fitted model and data used.}
\usage{
plot_fitted_model(fitted.model, ci)
}
\arguments{
\item{fitted.model}{List. Fitted model as created by the function \code{fit_model()}.}

\item{ci}{Numerical. Confidence interval of the fit displayed.}
}
\value{
A list of ggplot objects:
\itemize{
  \item \code{xy}: scattered plot of data and regression line.
  \item \code{timeseries}: fitted DAD time series in comparison with the actual data.
}
}
\description{
Plot fitted model and data used.
}
\examples{
set.seed(1234)

data.dad  = fake_dad()
data.expl = fake_expl(dad = data.dad)

prm = list(
 family = 'lm',
 data.dad = data.dad,
 data.expl = data.expl,
 varname.dad = 'count',
 varname.expl = 'thecount',
 date.range.fit = c('2020-01-10', '2020-12-15')
)

fitted.model = fit_model(prm)

g.fit = plot_fitted_model(fitted.model)
g.fit$xy
g.fit$timeseries

}
