% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_fitted_model_lm}
\alias{plot_fitted_model_lm}
\title{Plot the fitted model: regression line, data points used for the fit and fit metrics.}
\usage{
plot_fitted_model_lm(fitted_model)
}
\arguments{
\item{fitted_model}{List. Fitted model.}
}
\value{
A ggplot object.
}
\description{
Plot the fitted model: regression line, data points used for the fit and fit metrics.
}
