% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{process_newdata_expl}
\alias{process_newdata_expl}
\title{Process new explanatory data before nowcasting.}
\usage{
process_newdata_expl(newdata.expl, fitted.model)
}
\arguments{
\item{newdata.expl}{Dataframe of new explanatory data}

\item{fitted.model}{List. Fitted model.}
}
\value{
Processed dataframe
}
\description{
Process new explanatory data before nowcasting.
}
