% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{process_varname_dad}
\alias{process_varname_dad}
\title{Process variable name for DAD data}
\usage{
process_varname_dad(data.dad, varname.dad)
}
\arguments{
\item{data.dad}{Dataframe of DAD data}

\item{varname.dad}{String. Name of the variable in data.dad to fit the model to.}
}
\value{
Dataframe
}
\description{
Process variable name for DAD data
}
\keyword{internal}
