% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{process_varname_expl}
\alias{process_varname_expl}
\title{Process the variable name for explanatory data
by appending the suffix '.expl' to the variable name}
\usage{
process_varname_expl(data.expl, varname.expl)
}
\arguments{
\item{data.expl}{Dataframe of explanatory data}

\item{varname.expl}{String. Name of the variable in data.expl to use as an explanatory variable.}
}
\value{
Dataframe
}
\description{
Process the variable name for explanatory data
by appending the suffix '.expl' to the variable name
}
\keyword{internal}
