% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{remove_NAs}
\alias{remove_NAs}
\title{Remove all NA from the joint data set}
\usage{
remove_NAs(df, varname.dad, varname.expl)
}
\arguments{
\item{df}{Dataframe of joint data}

\item{varname.dad}{String. Variable used from the DAD data}

\item{varname.expl}{String. Variable used from the explanatory data}
}
\value{
Dataframe stripped of NAs
}
\description{
Remove all NA from the joint data set
}
\keyword{internal}
