% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data.R
\name{replace_zeros_by_tiny}
\alias{replace_zeros_by_tiny}
\title{Replace zeros by a tiny value (avoids log(0) downstream).}
\usage{
replace_zeros_by_tiny(df, varname.dad, varname.expl)
}
\arguments{
\item{df}{Dataframe of values.}

\item{varname.dad}{String. Name of the DAD variable whose zeros must be replaced by tiny value.}

\item{varname.expl}{String. Name of the explanatory variable whose zeros must be replaced by tiny value.}
}
\description{
Replace zeros by a tiny value (avoids log(0) downstream).
}
\keyword{internal}
