% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Fit a model to the data}
\usage{
fit_model(prm)
}
\arguments{
\item{prm}{List of parameters. Must contain:
 \itemize{
  \item \code{family} String. Type of model to fit. "lm" for linear regression.
"lm-log" for linear regression with log transformation.

  \item \code{data.dad} Dataframe of DAD data. Must contain the following columns:
date, virus, geo, count, percapita

  \item  \code{data.expl} Dataframe of explanatory data. Must contain the following columns:
date, virus, geo, varname.expl

  \item  \code{varname.dad} String. Name of the variable in data.dad to fit the model to.

  \item  \code{varname.expl} String. Name of the variable in data.expl to use as an explanatory variable.

  \item  \code{date.range.fit} Vector of two dates. The model will be fit to data within this date range.
 }}
}
\value{
A fitted model object.
}
\description{
Fit a model to the data
}
