% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nowcast.R
\name{nowcast}
\alias{nowcast}
\title{Nowcast DAD data using a fresh explanatory data set.}
\usage{
nowcast(fitted.model, newdata.expl, ci = 0.95, floor.zero = TRUE)
}
\arguments{
\item{fitted.model}{List. Fitted model.}

\item{newdata.expl}{Dataframe of new explanatory data.}

\item{ci}{Numerical. Confidence interval of the prediction. Must be between 0 and 1.}

\item{floor.zero}{Logical. If `TRUE` nowcast will be floored at 0 when estimates are negatives.}
}
\value{
Dataframe of nowcasted DAD values.
}
\description{
Nowcast DAD data using a fresh explanatory data set.
}
