% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_fitted_model}
\alias{plot_fitted_model}
\title{Plot fitted model and data used.}
\usage{
plot_fitted_model(fitted.model)
}
\arguments{
\item{fitted.model}{List. Fitted model as created by the function \code{fit_model()}.}
}
\value{
A ggplot object.
}
\description{
Plot fitted model and data used.
}
