% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_fitted_model}
\alias{plot_fitted_model}
\title{Plot fitted model and data used.}
\usage{
plot_fitted_model(fitted.model, ci)
}
\arguments{
\item{fitted.model}{List. Fitted model as created by the function \code{fit_model()}.}

\item{ci}{Numerical. Confidence interval of the fit displayed.}
}
\value{
A list of ggplot objects:
\itemize{
  \item \code{xy}: scattered plot of data and regression line.
  \item \code{timeseries}: fitted DAD time series in comparison with the actual data.
}
}
\description{
Plot fitted model and data used.
}
