% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_fitted_model_lm}
\alias{plot_fitted_model_lm}
\title{Plot the fitted model: regression line, data points used for the fit and fit metrics.}
\usage{
plot_fitted_model_lm(fitted.model)
}
\arguments{
\item{fitted.model}{List. Fitted model as created by the function \code{fit_model()}.}
}
\value{
A ggplot object.
}
\description{
Plot the fitted model: regression line, data points used for the fit and fit metrics.
}
\keyword{internal}
