% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_nowcast}
\alias{plot_nowcast}
\title{Plot the DAD nowcasting time series.}
\usage{
plot_nowcast(nowc)
}
\arguments{
\item{nowc}{List representing a nowcast object as returned
by the function \code{nowcast()}.}
}
\value{
A ggplot object.
}
\description{
Plot the nowcast DAD estimates along the new ("fresh")
explanatory data that fed the nowcast and the past
DAD and explanatory data that were used to fit the model.
}
