% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_timeseries_data}
\alias{plot_timeseries_data}
\title{Plot the time series of DAD and explanatory data used to fit the model.}
\usage{
plot_timeseries_data(fitted.model)
}
\arguments{
\item{fitted.model}{List. Fitted model as created by the function \code{fit_model()}}
}
\value{
A ggplot object.
}
\description{
Plot the time series of DAD and explanatory data used to fit the model.
}
\examples{

set.seed(1234)

data.dad  = fake_dad()
data.expl = fake_expl(dad = data.dad)

prm = list(
  family = 'lm',
  data.dad = data.dad,
  data.expl = data.expl,
  varname.dad = 'count',
  varname.expl = 'thecount',
  date.range.fit = c('2020-01-10', '2020-12-15')
)

fitted.model = fit_model(prm)

g = plot_timeseries_data(fitted.model)
g

}
